split scanner module
